[UI/#55] 기타 화면 및 로직 수정

+++ core/src/main/java/com/terning/core/designsystem/component/item/InternItem.kt
@@ -46,8 +46,8 @@ fun InternItem(
     workingPeriod: String,
     isScraped: Boolean,
     modifier: Modifier = Modifier,
-    scrapId: Int = 0,
-    onScrapButtonClicked: (Int) -> Unit = {},
+    scrapId: Long = 0,
+    onScrapButtonClicked: (Long) -> Unit = {},
 ) {
     Row(
         modifier = modifier
